dart-2-migration: Address incompatibilities

+++ exercises/difference-of-squares/lib/example.dart
@@ -1,9 +1,9 @@
 import "dart:math" show pow;
 
 class DifferenceOfSquares {
-  int squareOfSum(int input) => pow(_sum(input), 2);
+  int squareOfSum(int input) => pow(_sum(input), 2).toInt();
 
-  int sumOfSquare(int input) => _range(input).map((i) => pow(i, 2)).reduce((r, i) => r + i);
+  int sumOfSquare(int input) => _range(input).map((i) => pow(i, 2)).reduce((r, i) => r + i).toInt();
 
   int differenceOfSquares(int input) => squareOfSum(input) - sumOfSquare(input);
